Refactor: License -> SampleImages
