+++ src/axios/order/editor/SampleImages.js
@@ -1,0 +1,6 @@
+import { SAMPLE_LIST } from '../../../constants/api';
+import request from '../../axios';
+
+export const getSampleList = async () => {
+  return request.get(SAMPLE_LIST());
+};
