+++ src/component/order/editor/panels/panelItems/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-anonymous-default-export */
 import Customize from './Customize';
 import Uploads from './Uploads';
-import Images from './Images';
+import Images from './SampleImages';
 import Layers from './Layers';
 import AI from './AI';
 import Text from './Text';
